[Refactor] 컨벤션에 의한 위치변경
